requirements/pip.txt
# region Base packages

pip==18.*

# endregion

# region Django

django
djangorestframework
django-filter

# endregion

# region Miscellaneous

markdown
Pillow
Werkzeug

# endregion
